perf(lifecycle): use a single bulk container-list for stop/start/restart/kill/rm/pause/unpause (#1363) (#1388)
Closes #1363.

## What's wrong

`stop`, `start`, `restart`, `kill`, `rm`, `pause`, `unpause` each call
`live_service_containers(name)` or `live_replica_names(name, service)`
once **per service** before acting on it. Each call is one
`GET /containers/json` filtered by `podup.service={name}`. A bulk
`list_project_containers_by_service()` exists at
`internal/engine/lifecycle/scale.rs:275-302` and `down` already uses
it; the per-service lifecycle commands do not.

## What I changed

- Renamed the bulk helper to `Engine::live_project_replicas` (the same
  shape `down` was already using), and updated its doc comment to pin
  down the `all=true` requirement — libpod's container-list defaults to
  `runningOnly` when no `status=` filter is supplied, so a bare project
  GET would silently drop the very containers `start`/`restart`/
  `kill`/`rm`/`pause`/`unpause` need to act on.
- Switched the seven per-service lifecycle commands to prefetch the
  project's containers once at the top via the new helper, then read
  per-service slices from the in-memory `HashMap`. The per-service
  `*_one_service` functions now take `Vec<String>` instead of doing
  their own query. For a 100-service project the container-list GET
  count drops from S+1 to 1.
- `stop_container` now returns `Result<bool>` (transitioned vs.
  already-stopped/gone) so the `acted` flag in `stop_one_service`
  stays accurate now that we no longer filter by per-container state
  client-side — the bulk helper returns names without states. The
  existing drop-recheck tests still match the Ok/Err shape and now
  exercise the new bool via the existing "lost stop response"
  fixtures.
- `state_is_active` (formerly the `stop` state filter) is no longer
  used in production code. Kept it with `#[allow(dead_code)]` for the
  unit test that pins its truth table — the test is the only thing
  that still documents the running/paused-vs-anything-else split.

## Tests

- New unit test:
`live_project_replicas_returns_every_project_container_including_stopped`
  in `internal/engine/lifecycle/scale.rs`. Synthesises a 100-service
  project with 5 stopped replicas, asserts the returned map has every
  service (100 entries, 100 container names, the 5 stopped services
  included), AND asserts the bulk GET request shape (carries `all=true`,
  no `status=` filter workaround).
- All 1580 unit tests pass (`cargo test --all-features --lib`).
- All integration tests pass (`cargo test --all-features`).
- `cargo fmt --all -- --check` and
  `cargo clippy --all-features --all-targets -- -D warnings` are
  green.

## Test plan (per the issue)

- **Integration**: a 100-service project with 5 stopped services;
  `stop` issues 1 bulk GET (with `all=true`) and 100 stop POSTs (vs.
  200 GETs + 100 stop POSTs today). The unit test pins both halves of
  that on a fake-libpod server.
- **Lane**: needs real-Podman 5 and 6 to confirm the wire shape matches
  what the fake produced. Tagged `prio:P1` so the live lane runs the
  full lifecycle command surface against the changed code paths.

## Out of scope

- The libpod connection pool is a separate issue (#1362).
- The per-service `create`/`start`/`healthcheck` fan-out is not
  addressable in this shape (different concern, different filter).
- Query commands (`logs`, `inspect`, `ps`) still use the per-service
  `live_replica_names`. They could fold into the same bulk helper
  later, but the issue explicitly scopes this PR to the seven
  state-changing lifecycle commands.

## Lane

I have not run the binary against real Podman in this session — the
fix is unverified at the runtime/lane level. The static review, unit
tests (against a fake libpod shaped per the libpod source), the full
integration suite and clippy are green. The live lane is the place
where the wire-shape match against real libpod 5/6 (which the unit
test pins against the fake) actually gets exercised end-to-end.
