+++ internal/engine/lifecycle/down_label.rs
@@ -54,7 +54,7 @@ impl Engine {
 		// Without this every container here fell back to the per-label hash
 		// instead of the sequential identity colour `ps` and the compose-file
 		// `down` path use — the same class of defect fixed for `logs` (#1082).
-		let by_service = self.list_project_containers_by_service().await?;
+		let by_service = self.live_project_replicas().await?;
 		let mut service_names: Vec<String> = by_service.keys().cloned().collect();
 		service_names.sort();
 		crate::ui::set_services(&service_names);
